+++ src/commands/mod.rs
@@ -1,4 +1,4 @@
 pub mod monitor;
 pub mod opts;
 mod output;
-pub mod toggle_saved;
+pub mod toggle_liked;