+++ src/server/mod.rs
@@ -1,4 +1,4 @@
 pub mod dbus;
 pub mod grpc;
-pub mod saved_tracker;
+pub mod liked_tracker;
 pub mod service;