+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2022-04-29
+### Changed
+- Fixed polybar formatting.
+
 ## [0.2.0] - 2022-04-27
 ### Added
 - Added `added` and `removed` classes for when adding and removing songs from liked songs.